+++ src/test/java/shop/mtcoding/projoctbodykey/survey/SurveyControllerTest.java
@@ -7,19 +7,12 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import shop.mtcoding.projoctbodykey._core.utils.JwtUtil;
-import shop.mtcoding.projoctbodykey.choiceanswer.ChoiceAnswerRequest;
-import shop.mtcoding.projoctbodykey.meal.MealRequest;
-import shop.mtcoding.projoctbodykey.meal.MealResponse;
 import shop.mtcoding.projoctbodykey.user.User;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 
 @Transactional
@@ -45,150 +38,76 @@ public class SurveyControllerTest {
     }
     
     @Test
-    public void mealList_test() throws Exception {
+    public void surveyList_test() throws Exception {
         // given
         
         
         // when
         ResultActions actions = mvc.perform(
-                get("/api/meal/2024-05-19")
+                get("/api/survey")
                         .header("Authorization", "Bearer " + jwt)
         );
 
         // eye
+//        String respBody = actions.andReturn().getResponse().getContentAsString();
+//        System.out.println("respBody : " + respBody);
+
+        // then
+        actions.andExpect(jsonPath("$.status").value(200));
+        actions.andExpect(jsonPath("$.msg").value("성공"));
+        actions.andExpect(jsonPath("$.body.[0].id").value(1));
+        actions.andExpect(jsonPath("$.body.[0].title").value("설문조사1"));
+        actions.andExpect(jsonPath("$.body.[0].isAttend").value("참여가능"));
+        actions.andExpect(jsonPath("$.body.[0].progress").value("진행중"));
+        actions.andExpect(jsonPath("$.body.[0].questionCount").value(19));
+
+    }
+
+
+    @Test
+    public void surveyDetail_test() throws Exception {
+        // given
+
+
+        // when
+        ResultActions actions = mvc.perform(
+                get("/api/survey/1")
+                        .header("Authorization", "Bearer " + jwt)
+        );
+        //eye
+//        String respBody = actions.andReturn().getResponse().getContentAsString();
+//        System.out.println("respBody : " + respBody);
+
+        // then
+        actions.andExpect(jsonPath("$.status").value(200));
+        actions.andExpect(jsonPath("$.msg").value("성공"));
+        actions.andExpect(jsonPath("$.body.surveyId").value(1));
+        actions.andExpect(jsonPath("$.body.surveyTitle").value("설문조사1"));
+        actions.andExpect(jsonPath("$.body.questions[0].questionId").value("1"));
+        actions.andExpect(jsonPath("$.body.questions[0].questionTitle").value("하루 평균 수면 시간은 얼마나 되십니까?"));
+        actions.andExpect(jsonPath("$.body.questions[0].choices[0].choiceId").value("1"));
+        actions.andExpect(jsonPath("$.body.questions[0].choices[0].choiceTitle").value("6시간 미만"));
+        actions.andExpect(jsonPath("$.body.questions[0].choices[0].choiceNumber").value("1"));
+
+    }
+
+    @Test
+    public void surveyDetail_fail_test() throws Exception {
+        // given
+
+
+        // when
+        ResultActions actions = mvc.perform(
+                get("/api/survey/9")
+                        .header("Authorization", "Bearer " + jwt)
+        );
+        //eye
         String respBody = actions.andReturn().getResponse().getContentAsString();
         System.out.println("respBody : " + respBody);
 
         // then
-        actions.andExpect(jsonPath("$.status").value(200));
-        actions.andExpect(jsonPath("$.msg").value("성공"));
-        actions.andExpect(jsonPath("$.body.recommendCal").value(1244.1));
-        actions.andExpect(jsonPath("$.body.recommendCarbon").value(746.5));
-        actions.andExpect(jsonPath("$.body.recommendProtein").value(174.2));
-        actions.andExpect(jsonPath("$.body.recommendFat").value(273.7));
-        actions.andExpect(jsonPath("$.body.mealList[0].mealId").value("1"));
-        actions.andExpect(jsonPath("$.body.mealList[0].mealImg").value("이미지"));
-        actions.andExpect(jsonPath("$.body.mealList[0].eatTime").value("아침"));
-        actions.andExpect(jsonPath("$.body.mealList[0].foods[0].foodId").value(1));
-        actions.andExpect(jsonPath("$.body.mealList[0].foods[0].foodName").value("바나나"));
-        actions.andExpect(jsonPath("$.body.mealList[0].foods[0].carbo").value(22.84));
-        actions.andExpect(jsonPath("$.body.mealList[0].foods[0].protein").value(1.09));
-        actions.andExpect(jsonPath("$.body.mealList[0].foods[0].fat").value(0.33));
-        actions.andExpect(jsonPath("$.body.mealList[0].foods[0].kcal").value(88.0));
-        actions.andExpect(jsonPath("$.body.mealList[0].foods[0].gram").value(100));
-    }
-
-    @Test
-    public void save_test() throws Exception {
-        // given
-        List<MealRequest.SaveDTO.Food> foods =new ArrayList<>();
-        MealRequest.SaveDTO.Food food = new MealRequest.SaveDTO.Food();
-        food.setFoodId(1);
-        food.setQty(5);
-        foods.add(food);
-        MealRequest.SaveDTO reqDTO = new MealRequest.SaveDTO();
-        reqDTO.setEatTime("간식");
-        reqDTO.setMealImg("이미지");
-        reqDTO.setFoods(foods);
-
-        String reqBody = om.writeValueAsString(reqDTO);
-
-
-        // when
-        ResultActions actions = mvc.perform(
-                post("/api/meal/2024-05-19")
-                        .header("Authorization", "Bearer " + jwt)
-                        .content(reqBody)
-                        .contentType(MediaType.APPLICATION_JSON)
-        );
-
-        //eye
-        String respBody = actions.andReturn().getResponse().getContentAsString();
-        System.out.println("respBody : "+respBody);
-
-        // then
-        actions.andExpect(jsonPath("$.status").value(200));
-        actions.andExpect(jsonPath("$.msg").value("성공"));
-        actions.andExpect(jsonPath("$.body.mealId").value(3));
-        actions.andExpect(jsonPath("$.body.mealImg").value("이미지"));
-        actions.andExpect(jsonPath("$.body.foods[0].foodId").value(1));
-        actions.andExpect(jsonPath("$.body.foods[0].foodQuantity").value(5));
-    }
-
-    @Test
-    public void save_fail_test() throws Exception {
-        // given
-        List<MealRequest.SaveDTO.Food> foods =new ArrayList<>();
-        MealRequest.SaveDTO.Food food = new MealRequest.SaveDTO.Food();
-        food.setFoodId(500);
-        food.setQty(5);
-        foods.add(food);
-        MealRequest.SaveDTO reqDTO = new MealRequest.SaveDTO();
-        reqDTO.setEatTime("간식");
-        reqDTO.setMealImg("이미지");
-        reqDTO.setFoods(foods);
-
-        String reqBody = om.writeValueAsString(reqDTO);
-
-
-        // when
-        ResultActions actions = mvc.perform(
-                post("/api/meal/2024-05-19")
-                        .header("Authorization", "Bearer " + jwt)
-                        .content(reqBody)
-                        .contentType(MediaType.APPLICATION_JSON)
-        );
-
-        //eye
-        String respBody = actions.andReturn().getResponse().getContentAsString();
-        System.out.println("respBody : "+respBody);
-
-        // then
         actions.andExpect(jsonPath("$.status").value(404));
-        actions.andExpect(jsonPath("$.msg").value("찾으시는 음식이 없습니다"));
-        actions.andExpect(jsonPath("$.body").isEmpty());
-
-    }
-
-    @Test
-    public void delete_test() throws Exception {
-        // given
-
-
-        // when
-        ResultActions actions = mvc.perform(
-                delete("/api/meal/2024-05-19/1")
-                        .header("Authorization", "Bearer " + jwt)
-        );
-
-        //eye
-        String respBody = actions.andReturn().getResponse().getContentAsString();
-        System.out.println("respBody : "+respBody);
-
-        // then
-        actions.andExpect(jsonPath("$.status").value(200));
-        actions.andExpect(jsonPath("$.msg").value("성공"));
-        actions.andExpect(jsonPath("$.body").isEmpty());
-    }
-
-    @Test
-    public void delete_fail_test() throws Exception {
-        // given
-
-
-        // when
-        ResultActions actions = mvc.perform(
-                delete("/api/meal/2024-05-19/100")
-                        .header("Authorization", "Bearer " + jwt)
-        );
-
-        //eye
-        String respBody = actions.andReturn().getResponse().getContentAsString();
-        System.out.println("respBody : "+respBody);
-
-        // then
-        actions.andExpect(jsonPath("$.status").value(404));
-        actions.andExpect(jsonPath("$.msg").value("식단이 없습니다"));
+        actions.andExpect(jsonPath("$.msg").value("해당 설문조사를 찾을 수 없습니다"));
         actions.andExpect(jsonPath("$.body").isEmpty());
     }
 }
